[MTS/TC] [76127754] Specify bound element
We can now specify which element the tracking pixel success
and error events get bound to. Before, it had to be bound to
the body.

If we specify an element, and there is no client code to handle
the event, the event will bubble up to the body, so nothing
should break. An exception may be if the client code stops event
propagation, and an ancestor element expects to receive the
event. That scenario seems to be outside the responsibility of
the autotagging library, though.