[MTS/TL/HH][111418676] Fetch data using data attrs
https://www.pivotaltracker.com/story/show/111418676

We need to find an element's item, section (which we call the subgroup
inside the jquery-autotagging.js codebase), and value. This code change
is based off of Rent.com's JavaScript tracking code.

We may need to implement a more nuanced way of finding the clicked
element; the Rent.com code deals with the target's parent element when
deciding what the click target is.